pyproject.toml
[project]
name = "ai-engineering-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.67.0",
    "groq>=1.1.1",
    "ipykernel>=7.2.0",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
